requirements.txt
Flask==2.3.2
tensorflow==2.13.0          # if using .h5 model
Pillow==10.2.0
gunicorn
